Variable Explorer: Rename dicteditor.py to collectionseditor.py
