+++ spyderlib/widgets/varexp/collectionseditor.py
